Port to g_get_os_info()
This removes a lot of code and unexports two symbols.
