+++ libfwupd/fwupd-common-private.h
@@ -23,8 +23,6 @@ GVariant *
 fwupd_hash_kv_to_variant(GHashTable *hash) G_GNUC_NON_NULL(1);
 GHashTable *
 fwupd_variant_to_hash_kv(GVariant *dict) G_GNUC_NON_NULL(1);
-gchar *
-fwupd_build_user_agent_system(void);
 
 #ifdef HAVE_GIO_UNIX
 GUnixInputStream *
