+++ libfwupd/fwupd-common.c
@@ -22,11 +22,7 @@
 #include <sys/mman.h>
 #endif
 
-#include <locale.h>
 #include <string.h>
-#ifdef HAVE_UTSNAME_H
-#include <sys/utsname.h>
-#endif
 #include <json-glib/json-glib.h>
 
 /**
@@ -151,242 +147,6 @@ fwupd_checksum_get_best(GPtrArray *checksums)
 		}
 	}
 	return NULL;
-}
-
-static gchar *
-fwupd_get_os_release_filename(void)
-{
-#ifndef _WIN32
-	const gchar *hostdir = g_getenv("FWUPD_HOSTDIR");
-	const gchar *sysconfdir = g_getenv("FWUPD_SYSCONFDIR");
-	g_autofree gchar *fn1 = NULL;
-
-	if (hostdir == NULL)
-		hostdir = "/";
-
-	/* override */
-	if (sysconfdir != NULL) {
-		g_autofree gchar *fn2 = g_build_filename(hostdir, sysconfdir, "os-release", NULL);
-		if (g_file_test(fn2, G_FILE_TEST_EXISTS))
-			return g_steal_pointer(&fn2);
-	}
-
-	/* host locations */
-	if (g_strcmp0(sysconfdir, "/etc") != 0) {
-		g_autofree gchar *fn2 = g_build_filename(hostdir, "/etc/os-release", NULL);
-		if (g_file_test(fn2, G_FILE_TEST_EXISTS))
-			return g_steal_pointer(&fn2);
-	}
-	fn1 = g_build_filename(hostdir, "/usr/lib/os-release", NULL);
-	if (g_file_test(fn1, G_FILE_TEST_EXISTS))
-		return g_steal_pointer(&fn1);
-#endif
-	return NULL;
-}
-
-#ifdef HOST_MACHINE_SYSTEM_DARWIN
-static GHashTable *
-fwupd_get_os_release_darwin(GError **error)
-{
-	g_autofree gchar *stdout = NULL;
-	g_autofree gchar *sw_vers = g_find_program_in_path("sw_vers");
-	g_auto(GStrv) split = NULL;
-	g_autoptr(GHashTable) hash = g_hash_table_new_full(g_str_hash, g_str_equal, g_free, g_free);
-	struct {
-		const gchar *key;
-		const gchar *val;
-	} kvs[] = {{"ProductName:", "NAME"},
-		   {"ProductVersion:", "VERSION_ID"},
-		   {"BuildVersion:", "VARIANT_ID"},
-		   {NULL, NULL}};
-
-	g_return_val_if_fail(error == NULL || *error == NULL, NULL);
-
-	/* macOS */
-	if (sw_vers == NULL) {
-		g_set_error_literal(error, FWUPD_ERROR, FWUPD_ERROR_READ, "No os-release found");
-		return NULL;
-	}
-
-	/* parse in format:
-	 *
-	 *    ProductName:    Mac OS X
-	 *    ProductVersion: 10.14.6
-	 *    BuildVersion:   18G103
-	 */
-	if (!g_spawn_command_line_sync(sw_vers, &stdout, NULL, NULL, error))
-		return NULL;
-	split = g_strsplit(stdout, "\n", -1);
-	for (guint j = 0; split[j] != NULL; j++) {
-		for (guint i = 0; kvs[i].key != NULL; i++) {
-			if (g_str_has_prefix(split[j], kvs[i].key)) {
-				g_autofree gchar *tmp = g_strdup(split[j] + strlen(kvs[i].key));
-				g_hash_table_insert(hash,
-						    g_strdup(kvs[i].val),
-						    g_strdup(g_strstrip(tmp)));
-			}
-		}
-	}
-	g_hash_table_insert(hash, g_strdup("ID"), g_strdup("macos"));
-	return g_steal_pointer(&hash);
-}
-#endif
-
-/**
- * fwupd_get_os_release_full:
- * @filename: (nullable): optional filename to load
- * @error: (nullable): optional return location for an error
- *
- * Loads information from a defined system os-release file.
- *
- * Returns: (transfer container) (element-type utf8 utf8): keys from os-release
- *
- * Since: 1.8.8
- **/
-GHashTable *
-fwupd_get_os_release_full(const gchar *filename, GError **error)
-{
-	g_autofree gchar *buf = NULL;
-	g_autofree gchar *filename2 = g_strdup(filename);
-	g_auto(GStrv) lines = NULL;
-	g_autoptr(GHashTable) hash = NULL;
-
-	g_return_val_if_fail(error == NULL || *error == NULL, NULL);
-
-	hash = g_hash_table_new_full(g_str_hash, g_str_equal, g_free, g_free);
-	if (filename2 == NULL)
-		filename2 = fwupd_get_os_release_filename();
-#ifdef HOST_MACHINE_SYSTEM_DARWIN
-	if (filename2 == NULL)
-		return fwupd_get_os_release_darwin(error);
-#endif
-	if (filename2 == NULL) {
-#if defined(_WIN32)
-		/* TODO: Read the Windows version */
-		g_hash_table_insert(hash, g_strdup("OS"), g_strdup("Windows"));
-#elif defined(__NetBSD__)
-		g_hash_table_insert(hash, g_strdup("OS"), g_strdup("NetBSD"));
-#elif defined(__OpenBSD__)
-		g_hash_table_insert(hash, g_strdup("OS"), g_strdup("OpenBSD"));
-#endif
-		if (g_hash_table_size(hash) > 0)
-			return g_steal_pointer(&hash);
-		g_set_error_literal(error, FWUPD_ERROR, FWUPD_ERROR_READ, "No os-release found");
-		return NULL;
-	}
-
-	/* load each line */
-	if (!g_file_get_contents(filename2, &buf, NULL, error))
-		return NULL;
-	lines = g_strsplit(buf, "\n", -1);
-	for (guint i = 0; lines[i] != NULL; i++) {
-		gsize len, off = 0;
-		g_auto(GStrv) split = NULL;
-
-		/* split up into sections */
-		split = g_strsplit(lines[i], "=", 2);
-		if (g_strv_length(split) < 2)
-			continue;
-
-		/* remove double quotes if set both ends */
-		len = strlen(split[1]);
-		if (len == 0)
-			continue;
-		if (split[1][0] == '\"' && split[1][len - 1] == '\"') {
-			off++;
-			len -= 2;
-		}
-		g_hash_table_insert(hash, g_strdup(split[0]), g_strndup(split[1] + off, len));
-	}
-	return g_steal_pointer(&hash);
-}
-
-/**
- * fwupd_get_os_release:
- * @error: (nullable): optional return location for an error
- *
- * Loads information from the system os-release file.
- *
- * Returns: (transfer container) (element-type utf8 utf8): keys from os-release
- *
- * Since: 1.0.7
- **/
-GHashTable *
-fwupd_get_os_release(GError **error)
-{
-	g_return_val_if_fail(error == NULL || *error == NULL, NULL);
-	return fwupd_get_os_release_full(NULL, error);
-}
-
-static gchar *
-fwupd_build_user_agent_os_release(void)
-{
-	const gchar *keys[] = {"NAME", "VERSION_ID", "VARIANT", NULL};
-	g_autoptr(GHashTable) hash = NULL;
-	g_autoptr(GPtrArray) ids_os = g_ptr_array_new();
-
-	/* get all keys */
-	hash = fwupd_get_os_release(NULL);
-	if (hash == NULL)
-		return NULL;
-
-	/* create an array of the keys that exist */
-	for (guint i = 0; keys[i] != NULL; i++) {
-		const gchar *value = g_hash_table_lookup(hash, keys[i]);
-		if (value != NULL)
-			g_ptr_array_add(ids_os, (gpointer)value);
-	}
-	if (ids_os->len == 0)
-		return NULL;
-	g_ptr_array_add(ids_os, NULL);
-	return g_strjoinv(" ", (gchar **)ids_os->pdata);
-}
-
-/**
- * fwupd_build_user_agent_system: (skip):
- **/
-gchar *
-fwupd_build_user_agent_system(void)
-{
-#ifdef HAVE_UTSNAME_H
-	struct utsname name_tmp;
-#endif
-	g_autofree gchar *locale = NULL;
-	g_autofree gchar *os_release = NULL;
-	g_autoptr(GPtrArray) ids = g_ptr_array_new_with_free_func(g_free);
-
-	/* system, architecture and kernel, e.g. "Linux i686 4.14.5" */
-#ifdef HAVE_UTSNAME_H
-	memset(&name_tmp, 0, sizeof(struct utsname));
-	if (uname(&name_tmp) >= 0) {
-		g_ptr_array_add(ids,
-				g_strdup_printf("%s %s %s",
-						name_tmp.sysname,
-						name_tmp.machine,
-						name_tmp.release));
-	}
-#endif
-
-	/* current locale, e.g. "en-gb" */
-#ifdef HAVE_LC_MESSAGES
-	locale = g_strdup(setlocale(LC_MESSAGES, NULL));
-#endif
-	if (locale != NULL) {
-		g_strdelimit(locale, ".", '\0');
-		g_strdelimit(locale, "_", '-');
-		g_ptr_array_add(ids, g_steal_pointer(&locale));
-	}
-
-	/* OS release, e.g. "Fedora 27 Workstation" */
-	os_release = fwupd_build_user_agent_os_release();
-	if (os_release != NULL)
-		g_ptr_array_add(ids, g_steal_pointer(&os_release));
-
-	/* convert to string */
-	if (ids->len == 0)
-		return NULL;
-	g_ptr_array_add(ids, NULL);
-	return g_strjoinv("; ", (gchar **)ids->pdata);
 }
 
 #define FWUPD_GUID_NAMESPACE_DEFAULT   "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
